requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Mail
python-dotenv
email-validator
razorpay
Werkzeug
Jinja2
itsdangerous
click